+++ CHANGES.rst
@@ -103,6 +103,7 @@ Development
 - Skip reading WiFi sensors when WiFi is disabled. Thanks @thiasB!
 - Optimize reading the HX711. Don't ever use ``read_average()``. Instead, just
   ``read()`` the sensor ten times for computing the median.
+- Port LTE attach procedure from "Autonome Zelle". Thanks, @wtf!
 
 
 2019-08-19 0.6.0